// web/src/generated/prisma/models/StoreOrder.ts
export type StoreOrderSelect<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = runtime.Types.Extensions.GetSelect<{
  id?: boolean
  storeId?: boolean
  userId?: boolean
  facilityId?: boolean
  orderNum?: boolean
  pickupCode?: boolean
  isPaid?: boolean
  paidDate?: boolean
  checkoutAttributes?: boolean
  checkoutRef?: boolean
  paymentMethodId?: boolean
  paymentStatus?: boolean
  refundAmount?: boolean
  returnStatus?: boolean
  shippingMethodId?: boolean
  shippingAddress?: boolean
  shippingCost?: boolean
  shippingStatus?: boolean
  orderStatus?: boolean
  currency?: boolean
  discount?: boolean
  paymentCost?: boolean
  taxRate?: boolean
  orderTax?: boolean
  orderTotal?: boolean
  createdAt?: boolean
  updatedAt?: boolean
  Store?: boolean | Prisma.StoreDefaultArgs<ExtArgs>
  User?: boolean | Prisma.StoreOrder$UserArgs<ExtArgs>
  OrderNotes?: boolean | Prisma.StoreOrder$OrderNotesArgs<ExtArgs>
  OrderItems?: boolean | Prisma.StoreOrder$OrderItemsArgs<ExtArgs>
  OrderItemView?: boolean | Prisma.StoreOrder$OrderItemViewArgs<ExtArgs>
  ShippingMethod?: boolean | Prisma.ShippingMethodDefaultArgs<ExtArgs>
  Shipment?: boolean | Prisma.StoreOrder$ShipmentArgs<ExtArgs>
  PaymentMethod?: boolean | Prisma.StoreOrder$PaymentMethodArgs<ExtArgs>
  Rsvp?: boolean | Prisma.StoreOrder$RsvpArgs<ExtArgs>
  WaitList?: boolean | Prisma.StoreOrder$WaitListArgs<ExtArgs>
  StoreLedger?: boolean | Prisma.StoreOrder$StoreLedgerArgs<ExtArgs>
  customerCreditLedgers?: boolean | Prisma.StoreOrder$customerCreditLedgersArgs<ExtArgs>
  customerFiatLedgers?: boolean | Prisma.StoreOrder$customerFiatLedgersArgs<ExtArgs>
  _count?: boolean | Prisma.StoreOrderCountOutputTypeDefaultArgs<ExtArgs>
}, ExtArgs["result"]["storeOrder"]>
// ...
export type StoreOrderOmit<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = runtime.Types.Extensions.GetOmit<"id" | "storeId" | "userId" | "facilityId" | "orderNum" | "pickupCode" | "isPaid" | "paidDate" | "checkoutAttributes" | "checkoutRef" | "paymentMethodId" | "paymentStatus" | "refundAmount" | "returnStatus" | "shippingMethodId" | "shippingAddress" | "shippingCost" | "shippingStatus" | "orderStatus" | "currency" | "discount" | "paymentCost" | "taxRate" | "orderTax" | "orderTotal" | "createdAt" | "updatedAt", ExtArgs["result"]["storeOrder"]>
export type StoreOrderInclude<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  Store?: boolean | Prisma.StoreDefaultArgs<ExtArgs>
  User?: boolean | Prisma.StoreOrder$UserArgs<ExtArgs>
  OrderNotes?: boolean | Prisma.StoreOrder$OrderNotesArgs<ExtArgs>
  OrderItems?: boolean | Prisma.StoreOrder$OrderItemsArgs<ExtArgs>
  OrderItemView?: boolean | Prisma.StoreOrder$OrderItemViewArgs<ExtArgs>
  ShippingMethod?: boolean | Prisma.ShippingMethodDefaultArgs<ExtArgs>
  Shipment?: boolean | Prisma.StoreOrder$ShipmentArgs<ExtArgs>
  PaymentMethod?: boolean | Prisma.StoreOrder$PaymentMethodArgs<ExtArgs>
  Rsvp?: boolean | Prisma.StoreOrder$RsvpArgs<ExtArgs>
  WaitList?: boolean | Prisma.StoreOrder$WaitListArgs<ExtArgs>
  StoreLedger?: boolean | Prisma.StoreOrder$StoreLedgerArgs<ExtArgs>
  customerCreditLedgers?: boolean | Prisma.StoreOrder$customerCreditLedgersArgs<ExtArgs>
  customerFiatLedgers?: boolean | Prisma.StoreOrder$customerFiatLedgersArgs<ExtArgs>
  _count?: boolean | Prisma.StoreOrderCountOutputTypeDefaultArgs<ExtArgs>
}
// ...
export type $StoreOrderPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "StoreOrder"
  objects: {
    Store: Prisma.$StorePayload<ExtArgs>
    User: Prisma.$UserPayload<ExtArgs> | null
    OrderNotes: Prisma.$OrderNotePayload<ExtArgs>[]
    OrderItems: Prisma.$OrderItemPayload<ExtArgs>[]
    OrderItemView: Prisma.$orderitemviewPayload<ExtArgs>[]
    ShippingMethod: Prisma.$ShippingMethodPayload<ExtArgs>
    Shipment: Prisma.$ShipmentPayload<ExtArgs>[]
    PaymentMethod: Prisma.$PaymentMethodPayload<ExtArgs> | null
    Rsvp: Prisma.$RsvpPayload<ExtArgs>[]
    WaitList: Prisma.$WaitListPayload<ExtArgs> | null
    StoreLedger: Prisma.$StoreLedgerPayload<ExtArgs>[]
    customerCreditLedgers: Prisma.$CustomerCreditLedgerPayload<ExtArgs>[]
    customerFiatLedgers: Prisma.$CustomerFiatLedgerPayload<ExtArgs>[]
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    storeId: string
    userId: string | null
    facilityId: string | null
    orderNum: number | null
    pickupCode: string | null
    isPaid: boolean
    paidDate: bigint | null
    checkoutAttributes: string
    checkoutRef: string
    paymentMethodId: string | null
    paymentStatus: number
    refundAmount: runtime.Decimal
    returnStatus: number
    shippingMethodId: string
    shippingAddress: string
    shippingCost: runtime.Decimal
    shippingStatus: number
    orderStatus: number
    currency: string
    discount: runtime.Decimal
    paymentCost: runtime.Decimal
    taxRate: runtime.Decimal
    orderTax: runtime.Decimal
    orderTotal: runtime.Decimal
    createdAt: bigint
    updatedAt: bigint
  }, ExtArgs["result"]["storeOrder"]>
  composites: {}
}

export type StoreOrderGetPayload<S extends boolean | null | undefined | StoreOrderDefaultArgs> = runtime.Types.Result.GetResult<Prisma.$StoreOrderPayload, S>
// ...
export type StoreOrder$UserArgs<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  /**
   * Select specific fields to fetch from the User
   */
  select?: Prisma.UserSelect<ExtArgs> | null
  /**
   * Omit specific fields from the User
   */
  omit?: Prisma.UserOmit<ExtArgs> | null
  /**
   * Choose, which related nodes to fetch as well
   */
  include?: Prisma.UserInclude<ExtArgs> | null
  where?: Prisma.UserWhereInput
}

/**
 * StoreOrder.OrderNotes
 */
export type StoreOrder$OrderNotesArgs<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  /**
   * Select specific fields to fetch from the OrderNote
   */
  select?: Prisma.OrderNoteSelect<ExtArgs> | null
  /**
   * Omit specific fields from the OrderNote
   */
  omit?: Prisma.OrderNoteOmit<ExtArgs> | null
  /**
   * Choose, which related nodes to fetch as well
   */
  include?: Prisma.OrderNoteInclude<ExtArgs> | null
  where?: Prisma.OrderNoteWhereInput
  orderBy?: Prisma.OrderNoteOrderByWithRelationInput | Prisma.OrderNoteOrderByWithRelationInput[]
  cursor?: Prisma.OrderNoteWhereUniqueInput
  take?: number
  skip?: number
  distinct?: Prisma.OrderNoteScalarFieldEnum | Prisma.OrderNoteScalarFieldEnum[]
}

/**
 * StoreOrder.OrderItems
 */
export type StoreOrder$OrderItemsArgs<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  /**
   * Select specific fields to fetch from the OrderItem
   */
  select?: Prisma.OrderItemSelect<ExtArgs> | null
  /**
   * Omit specific fields from the OrderItem
   */
  omit?: Prisma.OrderItemOmit<ExtArgs> | null
  /**
   * Choose, which related nodes to fetch as well
   */
  include?: Prisma.OrderItemInclude<ExtArgs> | null
  where?: Prisma.OrderItemWhereInput
  orderBy?: Prisma.OrderItemOrderByWithRelationInput | Prisma.OrderItemOrderByWithRelationInput[]
  cursor?: Prisma.OrderItemWhereUniqueInput
  take?: number
  skip?: number
  distinct?: Prisma.OrderItemScalarFieldEnum | Prisma.OrderItemScalarFieldEnum[]
}

/**
 * StoreOrder.OrderItemView
 */
export type StoreOrder$OrderItemViewArgs<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  /**
   * Select specific fields to fetch from the orderitemview
   */
  select?: Prisma.orderitemviewSelect<ExtArgs> | null
  /**
   * Omit specific fields from the orderitemview
   */
  omit?: Prisma.orderitemviewOmit<ExtArgs> | null
  /**
   * Choose, which related nodes to fetch as well
   */
  include?: Prisma.orderitemviewInclude<ExtArgs> | null
  where?: Prisma.orderitemviewWhereInput
  orderBy?: Prisma.orderitemviewOrderByWithRelationInput | Prisma.orderitemviewOrderByWithRelationInput[]
  cursor?: Prisma.orderitemviewWhereUniqueInput
  take?: number
  skip?: number
  distinct?: Prisma.OrderitemviewScalarFieldEnum | Prisma.OrderitemviewScalarFieldEnum[]
}

/**
 * StoreOrder.Shipment
 */
export type StoreOrder$ShipmentArgs<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  /**
   * Select specific fields to fetch from the Shipment
   */
  select?: Prisma.ShipmentSelect<ExtArgs> | null
  /**
   * Omit specific fields from the Shipment
   */
  omit?: Prisma.ShipmentOmit<ExtArgs> | null
  /**
   * Choose, which related nodes to fetch as well
   */
  include?: Prisma.ShipmentInclude<ExtArgs> | null
  where?: Prisma.ShipmentWhereInput
  orderBy?: Prisma.ShipmentOrderByWithRelationInput | Prisma.ShipmentOrderByWithRelationInput[]
  cursor?: Prisma.ShipmentWhereUniqueInput
  take?: number
  skip?: number
  distinct?: Prisma.ShipmentScalarFieldEnum | Prisma.ShipmentScalarFieldEnum[]
}

/**
 * StoreOrder.PaymentMethod
 */
export type StoreOrder$PaymentMethodArgs<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  /**
   * Select specific fields to fetch from the PaymentMethod
   */
  select?: Prisma.PaymentMethodSelect<ExtArgs> | null
  /**
   * Omit specific fields from the PaymentMethod
   */
  omit?: Prisma.PaymentMethodOmit<ExtArgs> | null
  /**
   * Choose, which related nodes to fetch as well
   */
  include?: Prisma.PaymentMethodInclude<ExtArgs> | null
  where?: Prisma.PaymentMethodWhereInput
}

/**
 * StoreOrder.Rsvp
 */
export type StoreOrder$RsvpArgs<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  /**
   * Select specific fields to fetch from the Rsvp
   */
  select?: Prisma.RsvpSelect<ExtArgs> | null
  /**
   * Omit specific fields from the Rsvp
   */
  omit?: Prisma.RsvpOmit<ExtArgs> | null
  /**
   * Choose, which related nodes to fetch as well
   */
  include?: Prisma.RsvpInclude<ExtArgs> | null
  where?: Prisma.RsvpWhereInput
  orderBy?: Prisma.RsvpOrderByWithRelationInput | Prisma.RsvpOrderByWithRelationInput[]
  cursor?: Prisma.RsvpWhereUniqueInput
  take?: number
  skip?: number
  distinct?: Prisma.RsvpScalarFieldEnum | Prisma.RsvpScalarFieldEnum[]
}

/**
 * StoreOrder.WaitList
 */
export type StoreOrder$WaitListArgs<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  /**
   * Select specific fields to fetch from the WaitList
   */
  select?: Prisma.WaitListSelect<ExtArgs> | null
  /**
   * Omit specific fields from the WaitList
   */
  omit?: Prisma.WaitListOmit<ExtArgs> | null
  /**
   * Choose, which related nodes to fetch as well
   */
  include?: Prisma.WaitListInclude<ExtArgs> | null
  where?: Prisma.WaitListWhereInput
}

/**
 * StoreOrder.StoreLedger
 */
export type StoreOrder$StoreLedgerArgs<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  /**
   * Select specific fields to fetch from the StoreLedger
   */
  select?: Prisma.StoreLedgerSelect<ExtArgs> | null
  /**
   * Omit specific fields from the StoreLedger
   */
  omit?: Prisma.StoreLedgerOmit<ExtArgs> | null
  /**
   * Choose, which related nodes to fetch as well
   */
  include?: Prisma.StoreLedgerInclude<ExtArgs> | null
  where?: Prisma.StoreLedgerWhereInput
  orderBy?: Prisma.StoreLedgerOrderByWithRelationInput | Prisma.StoreLedgerOrderByWithRelationInput[]
  cursor?: Prisma.StoreLedgerWhereUniqueInput
  take?: number
  skip?: number
  distinct?: Prisma.StoreLedgerScalarFieldEnum | Prisma.StoreLedgerScalarFieldEnum[]
}
// ...
export type StoreOrderDefaultArgs<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  /**
   * Select specific fields to fetch from the StoreOrder
   */
  select?: Prisma.StoreOrderSelect<ExtArgs> | null
  /**
   * Omit specific fields from the StoreOrder
   */
  omit?: Prisma.StoreOrderOmit<ExtArgs> | null
  /**
   * Choose, which related nodes to fetch as well
   */
  include?: Prisma.StoreOrderInclude<ExtArgs> | null
}